backend/fastapibackend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==42.0.2
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
deep-translator==1.11.4
distro==1.9.0
dnspython==2.5.0
executing==2.0.1
fastapi==0.109.2
frozenlist==1.4.1
greenlet==3.0.3
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.0.352
langchain-community==0.0.19
langchain-core==0.1.22
langsmith==0.0.87
loguru==0.7.2
marshmallow==3.20.2
matplotlib-inline==0.1.6
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.7.2
packaging==23.2
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.2.0
pinecone-client==2.2.4
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
pypdf==4.0.1
pypdfium2==4.27.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.5.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
win32-setctime==1.1.0
yarl==1.9.4
